requirements.txt
absl-py==0.15.0
anyio==3.3.4
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
audioread==2.1.9
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1629908509068/work
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
click==7.1.2
cycler==0.10.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1627074852499/work
decorator==5.1.0
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
ffmpeg-python==0.2.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h11==0.12.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
htmlmin==0.1.12
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.3
ImageHash==4.2.1
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1630267471994/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1631291090039/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1632763767549/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1631590360471/work
jams==0.3.4
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1610146791023/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1633656206378/work
joblib==1.0.1
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1634694106991/work
jupyter==1.0.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter-console==6.4.0
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1631852756538/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1631590465624/work
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1621455677251/work
matplotlib==3.4.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mir-eval==0.6
missingno==0.5.0
mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1624941302223/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multimethod==1.6
nbclient==0.5.4
nbconvert==6.2.0
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
networkx==2.6.3
norbert==0.2.1
notebook==6.4.5
numba==0.53.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pandas-profiling==3.1.0
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
phik==0.12.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
pooch==1.5.2
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1622586138406/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1629903925368/work
protobuf==3.18.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pumpp==0.5.0
pyACA==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==19.0.2
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
spleeter==2.3.0
tangled-up-in-unicode==0.1.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow>=2.5.2
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado @ file:///home/conda/feedstock_root/build_artifacts/terminado_1631128159040/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
threadpoolctl==3.0.0
torch==1.9.1
torchaudio==0.9.0a0+a85b239
torchvision==0.10.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094706440/work
tqdm==4.62.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1630423529112/work
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.7
vggish-keras==0.0.18
visions==0.7.4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1605475534911/work
wrapt==1.12.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1633302054558/work
